Add support for MATLAB-like cells.
When working on a longish script, its common to have chunks that need to
be executed frequently. For example, a computation you are working on that
spans multiple lines, or a plot that you are trying to get just right.
When using an IPython notebook, you would put these chunks in cells.
MATLAB has a similar idea, where %% markers deliminate cells, which can
be executed with a single key press.

With this PR, you can define cells like this:

    print('Performing computation')
    a = 1 + 2
    b = a ** a

    ## Begin new cell

    print('Result of computation:')
    print(b)

Now you can execute the two cells independently.