Don't require 'word' in the metadata dict
Accidentally didn't put back 'word' into the result. This is a big
problem if the user's completion_metadata doesn't manually add 'word' to
the dict. Oops. Intentionally wrote it so the user may override 'word'
in the metadata dict if desired (e.g. to add an opening
parenthesis to a function call).