Refactor to make use of async
Since the recent jupyter_client now fully supports async, it is time to
refac refactor jupyter-vim to fully embrace Python's async
framework. This provides some neater solutions to some of our async
mechanics.

The design is now such that upon calling `:JupyterConnect`, the
JupyterMessenger class spawns a background thread that runs
`asyncio.eventloop.run_forever()`. This event loop can then be used to
run async functions inside this thread using
`async.run_coroutine_threadsafe`. In effect, you can run as much
functions concurrently as you want, without having to ever spawn
additional threads. Some prominent async functions are:

 - `JupyterMessenger._async_connect`: finds and connects to a kernel
 - `JupyterMessenger.get_next_msg`: waits and returns the first incoming
     message on a channel.
 - `JupyterMessenger.execute_with_reply`: sends off some code for
     execution and waits for the reply.

The driving part of this refactor is that
`JupyterMessenger.get_next_msg` support multiple dispatch. This means an
async function can loop through all the messages in a channel, discard
or use them as needed, without preventing others from accessing the same
messages. This greatly simplifies the monitor, as it no longer needs to
be informed when it is allowed for it to read a message. This also makes
waiting for a reply to a specific message much simpler and more robust,
as messages that arrive out of order are no longer a problem: the caller
simply keeps on reading all messages until the desired response arrives.

Regarding the monitor, this commit repurposes it as a debug tool that
echos all traffic on the shell, iopub and control channels. This is much
more useful than the hacky console it was before. The
`g:jupyter_monitor_console` variable has been removed. Instead there is
`:JupyterStartMonitor` which opens the monitor window and starts
monitoring, and `:JupyterStopMonitor` which closes the window and stops
monitoring.

The `VimMessenger` class has been refactored out since its main purpose,
keeping track of a `Sync` object, is no longer necessary in the new
async framework.

Since Python2 does not support async, support has been dropped and the
code related to `Pythonx` and `Pyeval` has been removed.